project name != directory name
hsdev misidentifies projects if multiple projects exist within a
directory and the directory name is not the same as the project name
(the 'cabal' builder file minus the cabal extension). For example:

   - emuwork
     |- src-emubase
        | emubase.cabal
	| ...
     |- src-main
        | ...

hsdev will blithely assume that the 'emubase' project is really named
'src-emubase'. This causes interesting interaction problems with
SublimeHaskell, which does the correct thing.